COLLISION DETECTION
We are able to detect when the player collides with the cube in the scene. After colliding with the cube in the scene, we log in console a message, "Collision"

Added hitbox and helper to visually see what is happening.

Need to add physics when colliding with the objects around the scene.